others/requirements_ocr.txt
# PaddleOCR dependencies
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy < 2.0
rapidfuzz
opencv-python
opencv-contrib-python
cython
Pillow
pyyaml
requests
albumentations == 1.4.10
albucore == 0.0.13
sentencepiece
yacs
seqeval
pypandoc
attrdict3
python_docx
tokenizers == 0.19.1
imagesize

# PaddleNLP dependencies
jieba
colorlog
colorama
seqeval
dill < 0.3.5
multiprocess <= 0.70.12.2
datasets >= 2.0.0
tqdm
paddlefsl
sentencepiece
huggingface_hub >= 0.19.2
onnx >= 1.10.0
protobuf >= 3.20.2 ; platform_system != 'Windows'
protobuf == 3.20.2 ; platform_system == 'Windows'
paddle2onnx
Flask-Babel
visualdl
fastapi
uvicorn
typer
rich
safetensors
tool_helpers ; platform_system == 'Linux'
aistudio-sdk >= 0.1.3
jinja2
regex